requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2020.12.5
charset-normalizer==3.3.2
Django==4.2.9
django-filter==23.5
djangorestframework==3.14.0
docopt==0.6.2
drf-yasg==1.21.7
#GDAL==3.4.3 pipwin
gunicorn
idna==2.10
inflection==0.5.0
Js2Py==0.74
Markdown==3.5.2
packaging==23.2
pipwin==0.5.2
psycopg2==2.9.9
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-decouple==3.8
pytz==2023.3.post1
PyYAML==5.3
six==1.16.0
soupsieve==2.
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0
vyper
eth-brownie
django_crontab
bitarray>=1.2.1
whitenoise